requirements.txt
ipykernel
pandas
scikit-learn
nltk
matplotlib
wordcloud
xgboost
PyYAML